Added a new method to Tfs called "ComponentPerturbs"
This method checks to see whether a Tfs component would perturb a
particle's motion.  Does this by checking if a component has non-zero
HKICK and VKICK and whether or not its integrated strengths are
non-zero (combined with >0 Lengths).  This may be a non-exhaustive list
of parameters to check (but sufficient for my needs), so update as
necessary.  Returns a list of parameters which do cause perturabtion,
and an empty list if none.