Switch to explicit relative imports in __init__.py
Previous syntax is bad form as risk of aliasing but also one day in the distant
future, we will move to python3, and this form of import will lead to
undesired behaviour as:

> All import forms not starting with . are interpreted as absolute imports. (PEP 0328)

i.e. it will not look in the __init__'s dir for the modules, and so
will always do the wrong thing.